requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.40.0
aiosqlite==0.20.0
watchdog==6.0.0
python-dotenv==1.0.1
edge-tts==6.1.12
spacy==3.8.3
python-multipart==0.0.19
aiofiles==24.1.0
numpy==2.2.0
